+++ lib/td_dd_web/controllers/grant_request_status_controller.ex
@@ -9,12 +9,15 @@ defmodule TdDdWeb.GrantRequestStatusController do
 
   action_fallback TdDdWeb.FallbackController
 
-  def create(conn, %{"grant_request_id" => id, "status" => status}) do
+  def create(conn, %{"grant_request_id" => id, "status" => status} = params) do
+    status_reason = Map.get(params, "reason")
+
     with claims <- conn.assigns[:current_resource],
-         request <- Requests.get_grant_request!(id),
+         request <- Requests.get_grant_request!(id, claims),
          {:can, true} <- {:can, can?(claims, approve(request))},
-         {:ok, _grant_request_status} <- Statuses.create_grant_request_status(request, status),
-         updated_request <- Requests.get_grant_request!(id) do
+         {:ok, _grant_request_status} <-
+           Statuses.create_grant_request_status(request, status, status_reason),
+         updated_request <- Requests.get_grant_request!(id, claims) do
       conn
       |> put_status(:created)
       |> put_view(GrantRequestView)