+++ priv/repo/migrations/20211007152427_add_explain_column_to_rule_results.exs
@@ -1,0 +1,11 @@
+defmodule TdDd.Repo.Migrations.AddExplainColumnToRuleResults do
+  use Ecto.Migration
+
+  def up do
+    alter(table("rule_results"), do: add(:details, :map, default: %{}))
+  end
+
+  def down do
+    alter(table("rule_results"), do: remove(:details))
+  end
+end